+++ assets/setup-tests.js
@@ -1,0 +1,3 @@
+require('@testing-library/jest-dom')
+require('@testing-library/react')
+require('regenerator-runtime/runtime')