+++ assets/src/components/activities/check_all_that_apply/CheckAllThatApplyDelivery.tsx
@@ -12,6 +12,7 @@ import { Stem } from '../common/DisplayedStem';
 import { Hints } from '../common/DisplayedHints';
 import { Reset } from '../common/Reset';
 import { Evaluation } from '../common/Evaluation';
+import { IconCorrect, IconIncorrect } from 'components/misc/Icons';
 
 type Evaluation = {
   score: number,
@@ -29,7 +30,7 @@ interface ChoicesProps {
 const Choices = ({ choices, selected, context, onSelect, isEvaluated }: ChoicesProps) => {
   const isSelected = (choiceId: ActivityTypes.ChoiceId) => !!selected.find(s => s === choiceId);
   return (
-    <div className="choices">
+    <div className="choices" aria-label="check all that apply choices">
       {choices.map((choice, index) =>
         <Choice
           key={choice.id}
@@ -54,6 +55,7 @@ interface ChoiceProps {
 const Choice = ({ choice, index, selected, context, onClick, isEvaluated }: ChoiceProps) => {
   return (
     <div key={choice.id}
+      aria-label={`choice ${index + 1}`}
       onClick={isEvaluated ? undefined : onClick}
       className={`choice ${selected ? 'selected' : ''}`}>
       <span className="choice-index">{index + 1}</span>
@@ -62,7 +64,8 @@ const Choice = ({ choice, index, selected, context, onClick, isEvaluated }: Choi
   );
 };
 
-const CheckAllThatApply = (props: DeliveryElementProps<CheckAllThatApplyModelSchema>) => {
+export const CheckAllThatApplyComponent = (props:
+  DeliveryElementProps<CheckAllThatApplyModelSchema>) => {
 
   const [model, setModel] = useState(props.model);
   const [attemptState, setAttemptState] = useState(props.state);
@@ -165,10 +168,22 @@ const CheckAllThatApply = (props: DeliveryElementProps<CheckAllThatApplyModelSch
     <Hints key="hints" onClick={onRequestHint} hints={hints}
       hasMoreHints={hasMoreHints} isEvaluated={isEvaluated} context={writerContext} />];
 
+  const gradedDetails = props.graded && props.progressState === 'in_review' ? [
+    evaluationSummary] : null;
+
+  const correctnessIcon = attemptState.score === 0 ? <IconIncorrect /> : <IconCorrect />;
+
+  const gradedPoints = props.graded && props.progressState === 'in_review' ? [
+    <div className="text-info font-italic">
+      {correctnessIcon}
+      <span>Points: </span><span>{attemptState.score + ' out of '
+        + attemptState.outOf}</span></div>] : null;
+
   const maybeSubmitButton = props.graded
     ? null
     : (
       <button
+        aria-label="submit"
         className="btn btn-primary mt-2 float-right" disabled={isEvaluated} onClick={onSubmit}>
         Submit
       </button>
@@ -179,11 +194,13 @@ const CheckAllThatApply = (props: DeliveryElementProps<CheckAllThatApplyModelSch
       <div className="activity-content">
         <div>
           <Stem stem={stem} context={writerContext} />
+          {gradedPoints}
           <Choices choices={choices} selected={selected}
             onSelect={onSelect} isEvaluated={isEvaluated} context={writerContext} />
           {maybeSubmitButton}
         </div>
         {ungradedDetails}
+        {gradedDetails}
       </div>
       {reset}
     </div>
@@ -193,7 +210,7 @@ const CheckAllThatApply = (props: DeliveryElementProps<CheckAllThatApplyModelSch
 // Defines the web component, a simple wrapper over our React component above
 export class CheckAllThatApplyDelivery extends DeliveryElement<CheckAllThatApplyModelSchema> {
   render(mountPoint: HTMLDivElement, props: DeliveryElementProps<CheckAllThatApplyModelSchema>) {
-    ReactDOM.render(<CheckAllThatApply {...props} />, mountPoint);
+    ReactDOM.render(<CheckAllThatApplyComponent {...props} />, mountPoint);
   }
 }