+++ assets/webpack.config.js
@@ -98,11 +98,22 @@ module.exports = (env, options) => ({
   module: {
     rules: [
       {
-        test: /\.js$/,
+        test: /\.js(x?)$/,
         exclude: /node_modules/,
         use: {
           loader: 'babel-loader'
         }
+      },
+      {
+        test: /\.ts(x?)$/, use: [
+          {
+            loader: 'babel-loader',
+            options: {
+              cacheDirectory: true
+            },
+          },
+          { loader: 'ts-loader' }
+        ], exclude: /node_modules/
       },
       {
         test: /\.[s]?css$/,
@@ -129,29 +140,7 @@ module.exports = (env, options) => ({
           }
         ],
       },
-      {
-        test: /\.jsx$/,
-        exclude: /node_modules/,
-        use: {
-          loader: 'babel-loader'
-        }
-      },
       { test: /\.(png|gif|jpg|jpeg|svg)$/, use: 'file-loader' },
-      { test: /\.ts$/, use: ['babel-loader', 'ts-loader'], exclude: /node_modules/ },
-      {
-        test: /\.tsx$/, use: [
-          {
-            loader: 'babel-loader',
-            options: {
-              // This is a feature of `babel-loader` for webpack (not Babel itself).
-              // It enables caching results in ./node_modules/.cache/babel-loader/
-              // directory for faster rebuilds.
-              cacheDirectory: true
-            },
-          },
-          { loader: 'ts-loader' }
-        ], exclude: /node_modules/
-      }
     ]
   },
   plugins: [